Create Alternate Adder
Creating an alternate adder to take the interpretation of every stage in
"The 12 Days of Christmas" as a list of all items received that day
(i.e. the end of the second day results in the subject having 4 items,
the partridge in the pear tree from the first day, one from the second
day, and the 2 french hens).